Published version: Robotic Welder Arm Control v1
