+++ MultigridProjectorPrograms/RobotArm/world_steam_description.txt
@@ -1,0 +1,30 @@
+[h1]Robotic Welder Arm Control v1[/h1]
+
+[h2]Requirements[/h2]
+[list]
+[*] Experimental mode enabled in the game (so the Programmable Block works)
+[*] [url=https://steamcommunity.com/sharedfiles/filedetails/?id=2415983416]Multigrid Projector client plugin[/url]
+[/list]
+
+[h2]How to use[/h2]
+[list]
+[*] Create and enter the world
+[*] Join the Testers faction
+[*] Spawn to the medical room
+[*] Enter the pilot seat at the top
+[*] Recompile the program in the [b]Shipyard Projector[/b]
+[/list]
+
+[url=https://steamcommunity.com/sharedfiles/filedetails/?id=2517216899]The script is here![/url]
+
+[h2]Remarks[/h2]
+[list]
+[*] Tested only in single player in creative mode so far.
+[*] Should work in multiplayer with simpler / less number of arms.
+[*] The video shows an earlier iteration with more arms. The new code works better with less arms.
+[*] The code may be refactored later to improve its usability in different worlds and for other use cases.
+[*] Works best with skinny blueprints, primarily small block ones. Solid large block ships are blocking out the arms.
+[*] Yes, you can build your own robot arms. The base must be in the "Welder Arms" group. Simpler is better.
+[/list]
+
+[i]Enjoy![/i]
